add new js file and import as a module
