+++ src/style.css
@@ -82,7 +82,7 @@ span {
   align-items: center;
   justify-content: space-between;
   border-bottom: 1px solid #c16ecf;
-  margin-bottom: 1px;
+  /* margin-bottom: 1px; */
   padding: 4% 2%;
   background: peachpuff;
 }
